requirements.txt
anyio==4.10.0
ariadne==0.26.2
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
graphql-core==3.2.5
grpcio==1.75.0
grpcio-tools==1.75.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
protobuf==6.32.1
py-mon==2.1.0
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
typing_extensions==4.15.0
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
